feat: refactor carnet fare contract
