+++ src/fare-contracts/FareContractInfo.tsx
@@ -24,8 +24,8 @@ import {
 import React from 'react';
 import {View} from 'react-native';
 import {
+  getLastUsedAccess,
   getValidityStatus,
-  isSentFareContract,
   isValidFareContract,
   mapToUserProfilesWithCount,
   useNonInspectableTokenWarning,
@@ -34,7 +34,6 @@ import {
   ValidityStatus,
 } from '../fare-contracts/utils';
 import {FareContractDetail} from '../fare-contracts/components/FareContractDetail';
-import {getLastUsedAccess} from './carnet/CarnetDetails';
 import {InspectionSymbol} from '../fare-contracts/components/InspectionSymbol';
 import {UserProfileWithCount} from './types';
 import {FareContractHarborStopPlaces} from './components/FareContractHarborStopPlaces';
@@ -147,7 +146,7 @@ export const FareContractInfoDetails = (
   );
 
   const isValidOrSentFareContract: boolean =
-    isValidFareContract(status) || isSentFareContract(status);
+    isValidFareContract(status) || status === 'sent';
 
   return (
     <View style={styles.container} accessible={true}>
@@ -167,10 +166,7 @@ export const FareContractInfoDetails = (
           )}
         </View>
         {isValidOrSentFareContract && (
-          <InspectionSymbol
-            {...props}
-            sentTicket={isSentFareContract(status)}
-          />
+          <InspectionSymbol {...props} sentTicket={status === 'sent'} />
         )}
       </View>
     </View>
