+++ src/fare-contracts/details/DetailsContent.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import {
   FareContract,
   isPreActivatedTravelRight,
+  isSentOrReceivedFareContract,
   NormalTravelRight,
 } from '@atb/ticketing';
 import {FareContractTexts, useTranslation} from '@atb/translations';
@@ -9,10 +10,8 @@ import {
   FareContractInfoHeader,
   FareContractInfoDetails,
 } from '../FareContractInfo';
-import {ValidityHeader} from '../ValidityHeader';
-import {ValidityLine} from '../ValidityLine';
 import {
-  getValidityStatus,
+  getFareContractInfo,
   mapToUserProfilesWithCount,
 } from '@atb/fare-contracts/utils';
 import {useMobileTokenContextState} from '@atb/mobile-token';
@@ -39,8 +38,13 @@ import {Barcode} from './Barcode';
 import {MapFilterType} from '@atb/components/map';
 import {MessageInfoText} from '@atb/components/message-info-text';
 import {useGetPhoneByAccountIdQuery} from '@atb/on-behalf-of/queries/use-get-phone-by-account-id-query';
+import {useAuthState} from '@atb/auth';
+import {UsedAccessValidityHeader} from '../carnet/UsedAccessValidityHeader';
+import {CarnetFooter} from '../carnet/CarnetFooter';
 import {MobilityBenefitsActionSectionItem} from '@atb/mobility/components/MobilityBenefitsActionSectionItem';
 import {useOperatorBenefitsForFareProduct} from '@atb/mobility/use-operator-benefits-for-fare-product';
+import {ValidityLine} from '../ValidityLine';
+import {ValidityHeader} from '../ValidityHeader';
 
 type Props = {
   fareContract: FareContract;
@@ -58,37 +62,47 @@ export const DetailsContent: React.FC<Props> = ({
   now,
   onReceiptNavigate,
   onNavigateToMap,
-  isSentFareContract = false,
 }) => {
+  const {abtCustomerId: currentUserId} = useAuthState();
+
   const {t} = useTranslation();
   const styles = useStyles();
   const {findGlobalMessages} = useGlobalMessagesState();
 
-  const firstTravelRight = fc.travelRights[0];
+  const {
+    isCarnetFareContract,
+    travelRights,
+    fareContractValidityStatus,
+    fareContractValidFrom,
+    fareContractValidTo,
+    carnetAccessStatus,
+    validityStatus,
+    validFrom,
+    validTo,
+    maximumNumberOfAccesses,
+    numberOfUsedAccesses,
+  } = getFareContractInfo(now, fc, currentUserId);
+
+  const isSentOrReceived = isSentOrReceivedFareContract(fc);
+  const isSent = isSentOrReceived && fc.customerAccountId !== currentUserId;
+  const isReceived = isSentOrReceived && fc.purchasedBy != currentUserId;
+
+  const firstTravelRight = travelRights[0];
   const {tariffZones, userProfiles} = useFirestoreConfiguration();
   const {deviceInspectionStatus, barcodeStatus} = useMobileTokenContextState();
   const {benefits} = useOperatorBenefitsForFareProduct(
     preassignedFareProduct?.id,
   );
 
-  const validityStatus = getValidityStatus(now, fc, isSentFareContract);
+  // If the ticket is received, get the sender account ID to look up for phone number.
+  const senderAccountId = isReceived ? fc.purchasedBy : undefined;
 
-  // Checks if the FareContract is purchased by a different ID,
-  // then if yes, return the purchaser ID, otherwise return blank.
-  const accountId =
-    !isSentFareContract && fc.purchasedBy !== fc.customerAccountId
-      ? fc.purchasedBy
-      : undefined;
+  const {data: purchaserPhoneNumber} =
+    useGetPhoneByAccountIdQuery(senderAccountId);
 
-  const {data: purchaserPhoneNumber} = useGetPhoneByAccountIdQuery(accountId);
-
-  if (isPreActivatedTravelRight(firstTravelRight)) {
-    const validFrom = firstTravelRight.startDateTime.toMillis();
-    const validTo = firstTravelRight.endDateTime.toMillis();
-
-    const {tariffZoneRefs} = firstTravelRight;
-    const firstZone = tariffZoneRefs?.[0];
-    const lastZone = tariffZoneRefs?.slice(-1)?.[0];
+  if (isPreActivatedTravelRight(firstTravelRight) || isCarnetFareContract) {
+    const firstZone = firstTravelRight.tariffZoneRefs?.[0];
+    const lastZone = firstTravelRight.tariffZoneRefs?.slice(-1)?.[0];
 
     const fromTariffZone = firstZone
       ? findReferenceDataById(tariffZones, firstZone)
@@ -116,13 +130,24 @@ export const DetailsContent: React.FC<Props> = ({
     return (
       <Section withBottomPadding>
         <GenericSectionItem>
-          <ValidityHeader
-            status={validityStatus}
-            now={now}
-            validFrom={validFrom}
-            validTo={validTo}
-            fareProductType={preassignedFareProduct?.type}
-          />
+          {isCarnetFareContract &&
+          fareContractValidityStatus === 'valid' &&
+          carnetAccessStatus ? (
+            <UsedAccessValidityHeader
+              now={now}
+              status={carnetAccessStatus}
+              validFrom={validFrom}
+              validTo={validTo}
+            />
+          ) : (
+            <ValidityHeader
+              status={fareContractValidityStatus}
+              now={now}
+              validFrom={fareContractValidFrom}
+              validTo={fareContractValidTo}
+              fareProductType={preassignedFareProduct?.type}
+            />
+          )}
           <ValidityLine
             status={validityStatus}
             now={now}
@@ -135,9 +160,7 @@ export const DetailsContent: React.FC<Props> = ({
             status={validityStatus}
             testID="details"
             preassignedFareProduct={preassignedFareProduct}
-            sentToCustomerAccountId={
-              isSentFareContract ? fc.customerAccountId : undefined
-            }
+            sentToCustomerAccountId={isSent ? fc.customerAccountId : undefined}
           />
         </GenericSectionItem>
         {deviceInspectionStatus === 'inspectable' &&
@@ -161,6 +184,15 @@ export const DetailsContent: React.FC<Props> = ({
             preassignedFareProduct={preassignedFareProduct}
           />
         </GenericSectionItem>
+        {isCarnetFareContract && (
+          <GenericSectionItem>
+            <CarnetFooter
+              active={validityStatus === 'valid'}
+              maximumNumberOfAccesses={maximumNumberOfAccesses!}
+              numberOfUsedAccesses={numberOfUsedAccesses!}
+            />
+          </GenericSectionItem>
+        )}
         {globalMessageCount > 0 && (
           <GenericSectionItem>
             <View style={styles.globalMessages}>
