+++ src/fare-contracts/index.ts
@@ -1,8 +1,6 @@
 export {FareContractAndReservationsList} from './FareContractAndReservationsList';
 export {FareContractHarborStopPlaces} from './components/FareContractHarborStopPlaces';
 export {SimpleFareContract} from './SimpleFareContract';
-export {CarnetDetailedView} from './carnet/CarnetDetailedView';
 export {DetailsContent} from './details/DetailsContent';
 export {getValidityStatus} from './utils';
-export {getLastUsedAccess} from './carnet/CarnetDetails';
 export type {UserProfileWithCount} from './types';
