+++ src/fare-contracts/utils.ts
@@ -1,9 +1,14 @@
 import {
+  CarnetTravelRight,
+  CarnetTravelRightUsedAccess,
   FareContract,
   FareContractState,
+  NormalTravelRight,
+  flattenCarnetTravelRightAccesses,
   isCarnet,
   isCarnetTravelRight,
   isPreActivatedTravelRight,
+  isSentOrReceivedFareContract,
 } from '@atb/ticketing';
 import {
   findReferenceDataById,
@@ -23,6 +28,7 @@ import {
 } from '@atb/translations';
 
 import {useMobileTokenContextState} from '@atb/mobile-token';
+import {UsedAccessStatus} from './carnet/types';
 
 export type RelativeValidityStatus = 'upcoming' | 'valid' | 'expired';
 
@@ -220,3 +226,110 @@ export const useDefaultPreassignedFareProduct = (
 
 export const getFareProductRef = (fc: FareContract) =>
   fc.travelRights[0]?.fareProductRef;
+
+type FareContractInfoProps = {
+  isCarnetFareContract: boolean;
+  travelRights: NormalTravelRight[];
+  fareContractValidityStatus: ValidityStatus;
+  fareContractValidFrom: number;
+  fareContractValidTo: number;
+  carnetAccessStatus?: UsedAccessStatus;
+  validityStatus: ValidityStatus;
+  validFrom: number;
+  validTo: number;
+  maximumNumberOfAccesses?: number;
+  numberOfUsedAccesses?: number;
+};
+
+export function getFareContractInfo(
+  now: number,
+  fc: FareContract,
+  currentUserId?: string,
+): FareContractInfoProps {
+  const isCarnetFareContract = isCarnet(fc);
+
+  const isSentOrReceived = isSentOrReceivedFareContract(fc);
+  const isSent = isSentOrReceived && fc.customerAccountId !== currentUserId;
+
+  const travelRights = fc.travelRights.filter(
+    isCarnetFareContract ? isCarnetTravelRight : isPreActivatedTravelRight,
+  );
+  const firstTravelRight = travelRights[0];
+
+  const fareContractValidFrom = firstTravelRight.startDateTime.toMillis();
+  const fareContractValidTo = firstTravelRight.endDateTime.toMillis();
+
+  const fareContractValidityStatus = getValidityStatus(now, fc, isSent);
+
+  const carnetTravelRightAccesses = isCarnetFareContract
+    ? flattenCarnetTravelRightAccesses(travelRights as CarnetTravelRight[])
+    : undefined;
+
+  const lastUsedAccess =
+    carnetTravelRightAccesses &&
+    getLastUsedAccess(now, carnetTravelRightAccesses.usedAccesses);
+
+  const validFrom = lastUsedAccess?.validFrom
+    ? lastUsedAccess.validFrom
+    : fareContractValidFrom;
+  const validTo = lastUsedAccess?.validTo
+    ? lastUsedAccess.validTo
+    : fareContractValidTo;
+  const validityStatus = lastUsedAccess
+    ? lastUsedAccess.status
+    : fareContractValidityStatus;
+
+  const carnetAccessStatus = lastUsedAccess?.status;
+  const maximumNumberOfAccesses =
+    carnetTravelRightAccesses?.maximumNumberOfAccesses;
+  const numberOfUsedAccesses = carnetTravelRightAccesses?.numberOfUsedAccesses;
+
+  return {
+    isCarnetFareContract,
+    travelRights,
+    fareContractValidityStatus,
+    fareContractValidFrom,
+    fareContractValidTo,
+    carnetAccessStatus,
+    validityStatus,
+    validFrom,
+    validTo,
+    maximumNumberOfAccesses,
+    numberOfUsedAccesses,
+  };
+}
+
+type LastUsedAccessState = {
+  status: UsedAccessStatus;
+  validFrom: number | undefined;
+  validTo: number | undefined;
+};
+
+function getUsedAccessValidity(
+  now: number,
+  validFrom: number,
+  validTo: number,
+): UsedAccessStatus {
+  if (now > validTo) return 'inactive';
+  if (now < validFrom) return 'upcoming';
+  return 'valid';
+}
+
+export function getLastUsedAccess(
+  now: number,
+  usedAccesses: CarnetTravelRightUsedAccess[],
+): LastUsedAccessState {
+  const lastUsedAccess = usedAccesses.slice(-1).pop();
+
+  let status: UsedAccessStatus = 'inactive';
+  let validFrom: number | undefined = undefined;
+  let validTo: number | undefined = undefined;
+
+  if (lastUsedAccess) {
+    validFrom = lastUsedAccess.startDateTime.toMillis();
+    validTo = lastUsedAccess.endDateTime.toMillis();
+    status = getUsedAccessValidity(now, validFrom, validTo);
+  }
+
+  return {status, validFrom, validTo};
+}
